import of the pure-Ansible tests
Drop the dependency against python-tripleo-helper and it's run_tests.
Use instead pure-Ansible tasks and roles.

We will now depend on a bunch of external roles (collect-logs,
validate-tempest, etc). At this point, those roles are not packages
don't have a stable branch.
We took the initiative to do a full import of these dependencies the
time the can improve the sitatuion.

Co-Author: Dimitri Savineau <dsavinea@redhat.com>

Change-Id: Id0855e2a858794289c3aaed59a4c1d135931401d