Use Openstack releasename
In order to make the agent compatible with RDO, we have to
use the same roles between similar topics (ie OSP10 is the
same than RDO-Newton).
That way we don't have to modify the playbooks to add a
new Openstack release because this will be provided by
a k/v in the topic.data field.

Change-Id: I489e6be8345f2deb6081453c687389059db3cbfb