upgrade: support topic without next_topic
If the topic has not next topic, or if the remoteci cannot access it
yet, the next dictionnary won't have any topic key.
This patch uses a variable to set the name of the next_topic_name, if
the next topic is not available, it still sets the variable but with the
'null' string. This way, the variable is a string and is always defined.

In addition, the patch avoid the user of job_definitions.topic, since
this key may not be set in the future (Ib8e03d5fd4c20def64d8d7442e154378e12063cc).

Change-Id: I8d8c5a14032e694eba1f1ed479f255f9a8a5acd9