run_tests: Configure director test correctly
In the certification test suite, the director test can be executed
from the undercloud but it requires to configure the
overcloud_node_ip paramater in the configuration.
We reuse the overcloud inventory from tripleo-inventory role to
select one of the node to run the director test.
In addition, because the rhcert command is executed as root user, we
need to configure the ssh passwordless authentification between
root@undercloud to heat-admin@overcloud node.

Change-Id: Idb60f1b6ff4b33fbfe0d09a45c38b1afb8a7660d
Depends-On: I619295b3b32bc90e5f2f7b208ce23d82b7b5d59a