registry: Use canonical project name
Currently the path of the images_list.txt is semi-hardcoded with some
regex in the middle to generate the path from dci_topic variable.
Because we already use the canonical project name from the component
to create the yum repository (which contain the file) we can reuse
the same variable.

Change-Id: I24892f991f62a6bbf036404d35a98867e9fbae23