registry: Add '-E' sudo flag to fetch_images.py
The fetch_images.py script gets the registry credentials from the env
variables DCI_REGISTRY_*. The user and password are stored in the
topics.data field on the API side.
In a CI context, we don't want to expose those credentials. If we
source a file with the DCI_REGISTRY_* variables before the agent run
the script will fail because the env variables from the user
(dci-ansible-agent) aren't keep when we switch to root user (via
become).

Change-Id: I1747c27eb91f9a835370fad7e56a5db47d9a56a2