fetch_bits: use the component of the current job
Until now, fetch_bits uses a "components" to know the components associated
with the job. Before this commit, the fact was set at the very beginning
of the job. This does not work when we do an update or an upgrade
because in this case a new job is created with some new components.

With this patch, we set "components" just before we use it.

Change-Id: Ia9bbb152ef07d06db0b89364992ee7f94a04c329