avoid a memory leak on the registry storage dir
If we remove all the tag of the existing image, we won't be able to actuallly
delete it from the registry. As a result, the registry size will grow up indefinitely.
Any tag can potentially be already used by an image and by the last "string"
attached to it. So before we apply a tag, we delete any potential existing tag.
See: https://github.com/docker/distribution/issues/1844

Change-Id: I43ab70660a68d230487124a270dfd635bda16469